+++ libs/integration-tests/src/lib/start.ts
@@ -1,0 +1,8 @@
+import { PlatformManager } from './platform/platform-manager'
+
+async function runAllContainers() {
+  const manager = new PlatformManager()
+  await manager.startContainers()
+}
+
+runAllContainers()